mock_vneid/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
python-jose[cryptography]>=3.3.0
jinja2>=3.1.0
